+++ app/mypage/_components/styles/myType.module.scss
