+++ app/mypage/_components/styles/settingNav.module.scss
