+++ app/mypage/_components/styles/type.module.scss
