+++ app/mypage/profileSetting/_component/CustomProfile.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import style from './profileSetting.module.scss'
+import style from './styles/profileSetting.module.scss'
 import { useEffect, useState } from 'react'
 import CustomResult from './CustomContainer'
